+++ packages/typography/text/index.ts
@@ -1,0 +1,1 @@
+export { default, TextProps, TextTags } from './Text';