+++ sb/stories/Typography.stories.tsx
@@ -4,7 +4,8 @@ import { ThemeProvider } from 'styled-components';
 import { theme } from '@atlantum/theme';
 import GlobalStyles from '@atlantum/theme';
 import './stories.styles.css';
-import Heading from '../../packages/typography/heading';
+import Heading from '@atlantum/heading';
+import Text from '@atlantum/text';
 
 const Typography = () => {
     return (
@@ -18,6 +19,16 @@ const Typography = () => {
                     <Heading as={'h4'}>Heading 4</Heading>
                     <Heading as={'h5'}>Heading 5</Heading>
                     <Heading as={'h6'}>Heading 6</Heading>
+                </div>
+
+                <div>
+                    <Text as={'p'}>
+                        Paragraph text <Text as={'i'}>Emphasize</Text>{' '}
+                        <Text as={'strong'}>Strong</Text>
+                    </Text>
+                    <Text as={'span'} size={'label'}>
+                        Span text
+                    </Text>
                 </div>
             </div>
         </ThemeProvider>