+++ packages/typography/heading/CHANGELOG.md
@@ -1,11 +1,12 @@
 # @atlantum/heading
 
 ## 0.1.0
+
 ### Minor Changes
 
-- f779f46: added Heading component, fixed token convention for typography, fixed input components
+-   f779f46: added Heading component, fixed token convention for typography, fixed input components
 
 ### Patch Changes
 
-- Updated dependencies [f779f46]
-  - @atlantum/theme@0.1.4
+-   Updated dependencies [f779f46]
+    -   @atlantum/theme@0.1.4